:art: Refactor URI processing by moving functions to a new utility module
